(MODULES-7444) Collapse V1/V2 adapter tests
 - With only a single adapter, the tests can be merged
